% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{accuracy}
\alias{accuracy}
\title{Calculate the accuracy of the prediction}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{An atomic vector, prediction results}

\item{truth}{An atomic vector, ground truth label}
}
\value{
A number
}
\description{
Calculate the accuracy of the prediction
}
