% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.r
\name{add}
\alias{add}
\title{Add together two numbers by Swati R.}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers by Swati R.
}
\examples{
add(1, 1)
}
