% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate the adjusted R-squared statistics}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{An atomic vector, prediction results}

\item{truth}{An atomic vector, ground truth label}
}
\value{
A number
}
\description{
Calculate the adjusted R-squared statistics
}
