% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{Find the classes of each variable in the input}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble.}
}
\value{
Same type as input x
}
\description{
Find the classes of each variable in the input
}
