% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{Scale the numeric columns of the input}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble.}

\item{center}{A boolean, whether centering}

\item{scale}{A boolean, whether scaling}
}
\value{
Same type as input x
}
\description{
Scale the numeric columns of the input
}
