% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{f1}
\alias{f1}
\title{Calculate the f1 score of the prediction}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{An atomic vector, prediction results}

\item{truth}{An atomic vector, ground truth label}
}
\value{
A number
}
\description{
Calculate the f1 score of the prediction
}
