% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{Check whether the element is the maximum}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector.}

\item{na.rm}{A boolean, whether na is allowed in input}
}
\value{
A logical vector
}
\description{
Check whether the element is the maximum
}
\examples{
is_max(c(0,1,1,0,3))
}
