% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Calculate the log-likelihood of x under the uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{min}{A number, min of x.}

\item{max}{A number, max of x.}
}
\value{
A number
}
\description{
Calculate the log-likelihood of x under the uniform distribution
}
