% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Find the maximum Bernoulli log-likelihood parameter}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector containing 0s and 1s representing Bernoulli trials.}
}
\value{
The value of p that maximizes the log-likelihood of the Bernoulli distribution for the given data.
}
\description{
This function calculates the log-likelihood of a Bernoulli distribution for a range of probability values (p) and then finds the p that maximizes the log-likelihood for the given data.
}
\details{
This function leverages the idea that a Bernoulli distribution is a special case of the Binomial distribution where the number of trials (n_trials) is 1. It calculates the log-likelihood using \code{dbinom} for a range of p values and then identifies the p that leads to the highest log-likelihood.
}
\examples{

# Simulate some Bernoulli trial data
data <- rbinom(n = 100, size = 1, prob = 0.6)

# Find the maximum log-likelihood parameter
estimated_p <- logLikBernoulli(data)

print(estimated_p)

}
