% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculate the minimum n needed for a two-sample t-test}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{A number, Cohen’s d.}

\item{power}{A number, using 0.,8 as the default}
}
\value{
A number
}
\description{
Calculate the minimum n needed for a two-sample t-test
}
