% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for T-Test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Numeric vector. One or two samples of preliminary data.}

\item{x2}{Numeric vector (optional). Second sample of preliminary data for a two-sample t-test.}

\item{power}{Numeric scalar. Desired statistical power (default is 0.8).}

\item{alpha}{Numeric scalar. Significance level (default is 0.05).}
}
\value{
Numeric scalar. Minimum sample size needed for the t-test.
}
\description{
This function calculates the minimum sample size needed for a t-test with specified power and significance level.
}
\details{
This function calculates the minimum sample size needed for a t-test, either one-sample or two-sample, with specified power and significance level. If only one sample is provided (\code{x1}), a one-sample t-test is assumed; otherwise, a two-sample t-test is assumed.
}
\examples{
# One-sample t-test
x <- rnorm(20)
minimumN(x)

# Two-sample t-test
x1 <- rnorm(20)
x2 <- rnorm(20)
minimumN(x1, x2)

}
