% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{precision}
\alias{precision}
\title{Calculate the precision of the prediction}
\usage{
precision(pred, truth)
}
\arguments{
\item{pred}{An atomic vector, prediction results}

\item{truth}{An atomic vector, ground truth label}
}
\value{
A number
}
\description{
Calculate the precision of the prediction
}
