% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Sampling from dataset x}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{An atomic vector or a dataframe-like object.}

\item{n}{A number}

\item{replace}{A boolean}
}
\value{
Same type as input x
}
\description{
Sampling from dataset x
}
\examples{
rando(c(1,2,3), 2)
}
