% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{recall}
\alias{recall}
\title{Calculate the recall of the prediction}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{An atomic vector, prediction results}

\item{truth}{An atomic vector, ground truth label}
}
\value{
A number
}
\description{
Calculate the recall of the prediction
}
