% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{Repeat the input data by rows or by columns}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A matrix or a dataframe.}

\item{M}{A integer, number of repeating by rows}

\item{N}{A integer, number of repeating by columns}
}
\value{
Same type as input x
}
\description{
Repeat the input data by rows or by columns
}
