% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize column names to small camel case}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A data frame object.}
}
\value{
A data frame with the same data but standardized column names in small camel case.
}
\description{
This function standardizes the column names of a data frame to small camel case (e.g., "columnA" becomes "columnA"). It utilizes \code{dplyr::rename_with} and \code{janitor::make_clean_names} to achieve consistent and valid variable names.
}
\details{
This function leverages the \code{dplyr} and \code{janitor} packages. It applies \code{janitor::make_clean_names} to each column name, cleaning and converting it to a valid R variable name in small camel case. Then, it uses \code{dplyr::rename_with} to perform the renaming based on the standardized names.
}
\examples{

# Example data frame
data <- data.frame(column_A = 1:5, Column_B = rnorm(5))

# Standardize column names to small camel case
standardized_data <- standardizeNames(data)

# Print the data frame with standardized names
head(standardized_data)

}
