% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survCurv}
\alias{survCurv}
\title{Plot a survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numeric vector indicating event status (1 = event, 0 = censored).}

\item{time}{A numeric vector of event times corresponding to the \code{status} vector.}
}
\value{
A ggplot object representing the Kaplan-Meier survival curve.
}
\description{
This function takes vectors of event times (\code{time}) and event indicators (\code{status}) and calculates the Kaplan-Meier (KM) survival curve. The KM curve estimates the probability of surviving beyond a specific time point for a population experiencing events over time.
}
\details{
This function utilizes dplyr verbs to manipulate the data and calculate the Kaplan-Meier curve elements. It first arranges the data by time. Then, it groups by time, calculates the number of events, censored observations, total at risk, hazard rate, and survival probability for each time point. Missing hazard rates (at the beginning) are replaced with 0. Finally, it accumulates survival probabilities, ungroups the data, adds an initial point (time=0, survival=1), rearranges by time, and selects the desired columns (time and survival) for the plot. The function then uses ggplot2 to create the survival curve visualization.
}
\examples{

# Simulate some survival data
set.seed(123)
time <- rexp(n = 20, rate = 0.1)
status <- rbinom(n = 20, size = 1, prob = 0.7)

# Generate Kaplan-Meier survival curve
km_curve <- survCurv(status, time)

# Print the ggplot object (use ggplot2::print() for detailed view)
km_curve

}
\references{
Kaplan, E. L., & Meier, P. (1958). Nonparametric estimation from incomplete observations. Journal of the American Statistical Association, 53(282), 457-481.
}
