% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{accuracy}
\alias{accuracy}
\title{Return the accuracy as the ratio of correct predictions to total predictions}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{A vector of predictions that are either logical or numeric, with
0 being false and 1 being true.}

\item{truth}{A vector of ground truths that are either logical or numeric, with
0 being false and 1 being true.}
}
\value{
The accuracy as a single number.
}
\description{
Return the accuracy as the ratio of correct predictions to total predictions
}
\examples{
pred1 <- c(TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE)
truth1 <- c(TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE)
accuracy(pred1, truth1)  # returns 0.57
}
