% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate the adjusted r-squared statistic between predicted and ground truth
values with a given number of parameterss n_p.}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{The predicted values (on a continuous scale).}

\item{truth}{The ground truth values (on a continuous scale).}

\item{n_p}{The number of model parameters.}
}
\value{
The adjusted r-squared statistic.
}
\description{
Calculate the adjusted r-squared statistic between predicted and ground truth
values with a given number of parameterss n_p.
}
\examples{
a <- c(6, 2, 7, 3, 9)
b <- c(3, 1, 7, 3, 6)
adj_R2(pred = a, truth = b, 3)  # returns -0.11
}
