% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{classes}
\alias{classes}
\title{Return a character vector containing the classes of each variable in the
given tibble.}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble.}
}
\value{
A vector containing the classes of each variable in the tibble.
}
\description{
Return a character vector containing the classes of each variable in the
given tibble.
}
\examples{
a <- tibble::tibble("a" = c(TRUE, FALSE), "b" = c(3, 4), "c" = c("Hi", "Hello"))
classes(a)  # returns   "logical"   "numeric" "character"
}
