% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{df_scale}
\alias{df_scale}
\title{Center and/or scale numeric columns of the given tibble.}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble.}

\item{center}{A logical indicating whether to center the numeric columns by
subtracting the mean.}

\item{scale}{A logical indicating whether to scale the numeric columns by
dividing by their standard deviations.}
}
\value{
A centered and/or scaled version of the input tibble.
}
\description{
Center and/or scale numeric columns of the given tibble.
}
\examples{
a <- tibble::tibble("a" = c(TRUE, FALSE, FALSE), "b" = c(1, 4, 5), "c" = c("Hi", "Hello", "What's up"))
df_scale(a)  # scales the "b" column so it becomes c(-1.12, 0.32, 0.801)
df_scale(a, scale = FALSE)  # centers the "b" column but doesn't center it
df_scale(a, center = FALSE)  # scales the "b" column but doesn't scale it
df_scale(a, scale = FALSE, center = FALSE)  # returns a unchanged
}
