% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download Report from REDCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable containing the REDCap API token.}

\item{redcapUrl}{The URL of the REDCap API.}

\item{redcapReportId}{The report ID number for the report you wish to download.}
}
\value{
A tibble containing the contents of the report.
}
\description{
Retrieves a specified report from REDCap using the API token, URL, and report ID provided.
The API token is sourced from the user's .Renviron file.
}
\examples{
# Example usage:
# downloadRedcapReport("REDCAP_API_TOKEN", "https://redcap.example.com/api/", "12345")

# Note: Before using the function, make sure that your .Renviron file contains a line like:
# REDCAP_API_TOKEN="your_api_token_here"
}
