% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Question 7: R function to download report from REDCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable containing the REDCap API token.}

\item{redcapUrl}{The URL of the REDCap API.}

\item{redcapReportId}{The report ID number for the report you wish to download.}
}
\value{
A tibble containing the contents of the report.
}
\description{
Retrieves a specified report from REDCap using the API token, URL, and report ID provided.
The API token is sourced from the user's .Renviron file.
}
\examples{

# downloadRedcapReport("REDCAP_API_TOKEN", "https://redcap.example.com/api/", "12345")

# Note: Before using the function, please make sure to follow following steps:
#1. Close R sessions
#2. Create a .Renviron file in your home directory (i.e., inside 'bmi510final' folder)
#3. Inside the .Renviron file, define the following environment variable:
#REDCAP_API_TOKEN=<your_api_token_here>
#4. Start your R session and now the R session should be able to load the new environment variables.
#5. Run the following code  
}
