% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{f1}
\alias{f1}
\title{Calculate the F1 score as the harmonic mean of precision and recall}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{A vector of predictions that are either logical or numeric, with
0 being false and 1 being true.}

\item{truth}{A vector of ground truths that are either logical or numeric, with
0 being false and 1 being true.}
}
\value{
The F1 score as a single number.
}
\description{
Calculate the F1 score as the harmonic mean of precision and recall
}
\examples{
pred1 <- c(TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE)
truth1 <- c(TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE)
f1(pred1, truth1)  # returns 0.67
}
