% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{is_min}
\alias{is_min}
\title{Return a logical vector that is TRUE at indices that have the minimum value
in the given vector.}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector.}

\item{na.rm}{A logical indicating whether to remove missing values.}
}
\value{
A logical vector of the same length as x that is TRUE at indices that
have the minimum value in the given vector x.
}
\description{
Return a logical vector that is TRUE at indices that have the minimum value
in the given vector.
}
\examples{
x <- c(3, 4, 8, 3, 5)
is_min(x)  # returns c(TRUE, FALSE, FALSE, TRUE, FALSE)
y <- c(3, NaN, 4, 6, 1)
is_min(y)  # returns c(FALSE, NA, FALSE, FALSE, TRUE)
is_min(y, FALSE)  # returns c(NA, NA, NA, NA, NA)
}
