% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Return the log-likelihood of the given sample under a chi-squared distribution
with the given degrees of freedom.}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A numeric vector of samples.}

\item{df}{Degrees of freedom (non-negative, but can be non-integer).}
}
\value{
The log-likelihood of the given sample under a chi-squared distribution
with the given degrees of freedom.
}
\description{
Return the log-likelihood of the given sample under a chi-squared distribution
with the given degrees of freedom.
}
\examples{
a <- stats::rchisq(100, 33)  # generate 100 samples from a chi-squared distribution with df=33
log_likelihood_chisq(a, 33)
}
