% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Return the log-likelihood of the given sample under an F distribution
with the given degrees of freedom df1 and df2.}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector of samples.}

\item{df1}{Degrees of freedom. Inf is allowed.}

\item{df2}{Degrees of freedom. Inf is allowed.}
}
\value{
The log-likelihood of the given sample under an F distribution
with the given degrees of freedom df1 and df2.
}
\description{
Return the log-likelihood of the given sample under an F distribution
with the given degrees of freedom df1 and df2.
}
\examples{
a <- stats::rf(100, 33, 40)  # generate 100 samples from an F distribution with df1=33 and df2=40
log_likelihood_f(a, 33, 40)
}
