% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Return the log-likelihood of the given sample under a normal distribution
with the given mean and standard deviation.}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector of samples.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}
}
\value{
The log-likelihood of the given sample under a normal distribution
with the given mean and standard deviation.
}
\description{
Return the log-likelihood of the given sample under a normal distribution
with the given mean and standard deviation.
}
\examples{
a <- stats::rnorm(100, 1, 3)  # generate 100 samples from a normal distribution with mean=1, sd=3
log_likelihood_norm(a, 1, 3)
}
