% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Return the log-likelihood of the given sample under a t distribution
with the given degrees of freedom.}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector of samples.}

\item{df}{Degrees of freedom (non-negative, but can be non-integer, Inf allowed).}
}
\value{
The log-likelihood of the given sample under a t distribution
with the given degrees of freedom.
}
\description{
Return the log-likelihood of the given sample under a t distribution
with the given degrees of freedom.
}
\examples{
a <- stats::rt(100, 33)  # generate 100 samples from a t distribution with df=33
log_likelihood_t(a, 33)
}
