% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Return the log-likelihood of the given sample under a uniform distribution
with the given minimum and maximum.}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector of samples.}

\item{min}{The minimum of the uniform distribution.}

\item{max}{The maximum of the uniform distribution.}
}
\value{
The log-likelihood of the given sample under a uniform distribution
with the given minimum and maximum.
}
\description{
Return the log-likelihood of the given sample under a uniform distribution
with the given minimum and maximum.
}
\examples{
a <- stats::runif(100, 1, 1000)  # generate 100 samples from a uniform distribution on the interval [1,1000]
log_likelihood_unif(a, 1, 1000)
}
