% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Question 1: R function to Maximize log-likelihood for Bernoulli distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{Numeric vector containing 0s and 1s.}
}
\value{
The parameter p that maximizes the log-likelihood.
}
\description{
This function computes the log-likelihood of the Bernoulli distribution parameter p
for a given binary data vector. It performs a grid-based search over possible values of
from 0 to 1 in steps of 0.001 to find the value that maximizes the log-likelihood of the data.
This approach provides an estimate for the parameter that is most likely to have generated
the observed data under a Bernoulli distribution assumption.
}
\examples{
print(logLikBernoulli(c(1, 0, 0, 0, 1, 1, 1)))
}
