% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Question 1: R function to Maximize log-likelihood for Bernoulli distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{Numeric vector containing 0s and 1s.}
}
\value{
The parameter p that maximizes the log-likelihood.
}
\description{
Question 1: R function to Maximize log-likelihood for Bernoulli distribution
}
