% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Return the minimum sample size needed for a two-sample t-test with the given
Cohen's d value and power.}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{true difference in means (Cohen's d)}

\item{power}{power of test (1 minus Type II error probability)}
}
\value{
The minimum sample size needed for a two-sample t-test with the given
Cohen's d value and power level.
}
\description{
Return the minimum sample size needed for a two-sample t-test with the given
Cohen's d value and power.
}
\examples{
minimum_n_per_group(1)  # returns 17
minimum_n_per_group(d = 0.01, power = 0.9)  # returns 210150
}
