% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{minimumN}
\alias{minimumN}
\title{Question 6: R function to calculate min. sample size for T-Test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Numeric vector of preliminary data for the first sample or the only sample.}

\item{x2}{Optional numeric vector of preliminary data for the second sample.}
}
\value{
An integer representing the minimum sample size needed to achieve the
specified power for a t-test.
}
\description{
This function estimates the minimum sample size required for a t-test with
80\% power and a significance level of 0.05. It can perform calculations
for both one-sample and two-sample t-tests based on the input provided.
}
\examples{
sample_size_one = minimumN(rnorm(30, mean = 5, sd = 1.5))
print(paste("Minimum sample size for one-sample t-test:", sample_size_one))
sample_size_two = minimumN(rnorm(30, mean = 5, sd = 1.5), rnorm(30, mean = 5.5, sd = 1.5))
print(paste("Minimum sample size for two-sample t-test:", sample_size_two))
}
