% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{pcApprox}
\alias{pcApprox}
\title{Principal Component Approximation of Data}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{npc}{Number of principal components to use.}
}
\value{
An approximation of the original data using the specified number of PCs.
}
\description{
Principal Component Approximation of Data
}
