% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{precision}
\alias{precision}
\title{Calculate precision as the ratio of true positives to total positive predictions; also
referred to as Positive Predictive Value (PPV).}
\usage{
precision(pred, truth)
}
\arguments{
\item{pred}{A vector of predictions that are either logical or numeric, with
0 being false and 1 being true.}

\item{truth}{A vector of ground truths that are either logical or numeric, with
0 being false and 1 being true.}
}
\value{
The precision as a single number.
}
\description{
Calculate precision as the ratio of true positives to total positive predictions; also
referred to as Positive Predictive Value (PPV).
}
\examples{
pred1 <- c(TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE)
truth1 <- c(TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE)
precision(pred1, truth1)  # returns 0.75
}
