% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{r2}
\alias{r2}
\title{Calculated the r-squared statistic between predicted and ground truth values.}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{The predicted values (on a continuous scale).}

\item{truth}{The ground truth values (on a continuous scale).}
}
\value{
The r-squared statistic between predicted and ground truth values.
}
\description{
Calculated the r-squared statistic between predicted and ground truth values.
}
\examples{
a <- c(6, 2, 7, 3, 9)
b <- c(3, 1, 7, 3, 6)
r2(pred = a, truth = b)  # returns 0.72
}
