% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{rando}
\alias{rando}
\title{Randomly sample items from a vector or rows from a dataframe-like object.}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{A vector or dataframe-like object.}

\item{n}{An integer number of samples.}

\item{replace}{Whether to sample with replacement.}
}
\value{
Items or rows sampled from the input object.
}
\description{
Randomly sample items from a vector or rows from a dataframe-like object.
}
\examples{
x <- 1:12
# get a random number from [1,12]
rando(x)
# get 5 random numbers from [1,12] without replacement
rando(x, 12, FALSE)
a <- data.frame("a" = c(1, 2, 3), "b" = c(4, 6, 8), "c" = c(10, 44, 32))
# sample 2 rows with replacement
rando(a, 2)
}
