% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{recall}
\alias{recall}
\title{Calculate recall as the ratio of true positives to total ground truth
positives (true positives + false negatives).}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{A vector of predictions that are either logical or numeric, with
0 being false and 1 being true.}

\item{truth}{A vector of ground truths that are either logical or numeric, with
0 being false and 1 being true.}
}
\value{
The recall as a single number.
}
\description{
Calculate recall as the ratio of true positives to total ground truth
positives (true positives + false negatives).
}
\examples{
pred1 <- c(TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE)
truth1 <- c(TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE)
recall(pred1, truth1)  # returns 0.6
}
