% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510final.R
\name{rep_mat}
\alias{rep_mat}
\title{Replicate the given matrix or dataframe a given number of times along rows
and/or columns.}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A matrix or dataframe to replicate.}

\item{M}{The integer number of times to replicate the input along rows.}

\item{N}{The integer number of times to replicate the input along columns}
}
\value{
A matrix with the replicated input matrix/dataframe
}
\description{
Replicate the given matrix or dataframe a given number of times along rows
and/or columns.
}
\examples{
x <- rbind(c(1,2), c(3,4))
rep_mat(x, 3, 1)  # replicates the matrix x 3 times row-wise
}
