% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{standardizeNames}
\alias{standardizeNames}
\title{Question 5: R function to standardize variable names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A data frame whose column names need to be standardized.}
}
\value{
A data frame with standardized column names in small camel case.
}
\description{
Standardize Column Names
}
\details{
This function standardizes the column names of a data frame to small camel case using
\code{janitor::make_clean_names} for initial cleaning and \code{snakecase::to_any_case}
for converting to the specified case format.
}
\examples{
# Sample data
sample_data = data.frame(
  `FiRst nAMe` = c("Swati", "Taylor"),
  `Last Name` = c("Rajwal", "Swift"),
  `Age in yeARS` = c(21, 32)
)
# Standardize names and print the result
print(standardizeNames(sample_data))
}
