% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardizes variable names in a data frame to small camel case}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble or data frame.}
}
\value{
A tibble with standardized variable names.
}
\description{
Standardizes variable names in a data frame to small camel case
}
