% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{survCurv}
\alias{survCurv}
\title{Calculates and plots a survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numerical vector indicating the status (1 = event, 0 = censored).}

\item{time}{Numerical vector of times at which the status is observed.}
}
\value{
Plot of the survival curve.
}
\description{
Calculates and plots a survival curve S(t)
}
