% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{survCurv}
\alias{survCurv}
\title{Question 2: R function to calculate and plot a survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numerical vector indicating the status (1 = event, 0 = censored).}

\item{time}{Numerical vector of times at which the status is observed.}
}
\value{
Plot of the survival curve.
}
\description{
This function calculates and plots a survival curve using time-to-event data.
It takes vectors of event statuses and times, computes the number of events
and the number at risk at each time point, and then uses these to calculate
the survival probability over time. The survival probability is plotted against
time to visualize the survival curve.
}
\examples{
# Load the survival dataset
survival_url = "https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv"
survival_data = read.csv(survival_url)
# Calculate and plot the survival curve
survCurv(survival_data$status, survival_data$time)
}
