% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{survCurv}
\alias{survCurv}
\title{Question 2: R function to calculate and plot a survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numerical vector indicating the status (1 = event, 0 = censored).}

\item{time}{Numerical vector of times at which the status is observed.}
}
\value{
Plot of the survival curve.
}
\description{
Question 2: R function to calculate and plot a survival curve S(t)
}
\examples{
survival_data = read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
survCurv(survival_data$status, survival_data$time)
}
