% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survCurv}
\alias{survCurv}
\title{Plot a survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numeric vector indicating event status (1 = event, 0 = censored).}

\item{time}{A numeric vector of event times corresponding to the \code{status} vector.}
}
\value{
A ggplot object representing the survival curve.
}
\description{
This function takes vectors of event times (\code{time}) and event indicators (\code{status}) and calculates the survival curve. The sur curve estimates the probability of surviving beyond a specific time point for a population experiencing events over time.
}
