% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{unscale}
\alias{unscale}
\title{Reverses the scaling and centering transformation applied to a vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled numeric vector.}
}
\value{
The unscaled original vector.
}
\description{
Reverses the scaling and centering transformation applied to a vector
}
