% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{unscale}
\alias{unscale}
\title{Question 3: R function to reverse the scaling and centering transformation applied to a vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled numeric vector.}
}
\value{
The unscaled original vector.
}
\description{
Question 3: R function to reverse the scaling and centering transformation applied to a vector
}
\examples{
scaled_vector = scale(c(10, 20, 30, 40, 50))
print(paste("Original data: ",scaled_vector))
print(paste("Original data after scaling: ",scaled_vector))
print(paste("Unscaling the scaled data:", unscale(scaled_vector))
}
